refactor(repl): factor out ReplEditor and ReplSession (#13131)
